stuck at expenses page UI rendering

+++ src/app/(routes)/dashboard/incomes/_components/IncomeList.jsx
@@ -18,7 +18,7 @@ function IncomeList() {
     const result = await db
       .select({
         ...getTableColumns(Incomes),
-        totalSpend: sql`sum((${Expenses.amount})::numeric)`.mapWith(Number),
+        totalSpend: sql`sum((${Expenses.amount}))`.mapWith(Number),
         totalItem: sql`count(${Expenses.id})`.mapWith(Number),
       })
       .from(Incomes)
